Refactoring

+++ afl_unicorn_plugin.py
@@ -41,34 +41,40 @@ class AflUnicornRunner(BackgroundTaskThread):
         outputs = DirectoryNameField('Select outputs folder')
         harness_file = OpenFileNameField('Select harness test file')
         get_form_input([separator, afl_binary, dumped_memory, inputs, outputs, harness_file], "Afl-unicorn Fuzzing Menu")
+
+        if(inputs.result == None or afl_binary.result == None or dumped_memory.result == None or outputs.result == None or harness_file.result == None):
+            return
+
+        if(len(afl_binary.result) <=0 or len(dumped_memory.result) <= 0 or len(inputs.result) <= 0 or len(outputs.result) <= 0 or len(harness_file.result) <=0):
+            show_message_box("Afl-Unicorn", "All fields are required !",
+                                            MessageBoxButtonSet.OKButtonSet, MessageBoxIcon.ErrorIcon)
+            return
         binja.log_info("Selected afl-fuzz binary: {0}".format(afl_binary.result))
         binja.log_info("Selected dumped memory folder: {0}".format(dumped_memory.result))
         binja.log_info("Selected inputs folder: {0}".format(inputs.result))
         binja.log_info("Selected outputs folder: {0}".format(outputs.result))
         binja.log_info("Selected harness test file: {0}".format(harness_file.result))
 
-        try:
-            if(len(afl_binary.result) <=0 or len(dumped_memory.result) <= 0 or len(inputs.result) <= 0 or len(outputs.result) <= 0 or len(harness_file.result) <=0):
-                show_message_box("Afl-Unicorn", "All fields are required !",
-                                                MessageBoxButtonSet.OKButtonSet, MessageBoxIcon.ErrorIcon)
-                return
-            self.afl_binary = afl_binary.result
-            self.dumped_memory = dumped_memory.result
-            self.inputs = inputs.result
-            self.outputs = outputs.result
-            self.harness_file = harness_file.result
+        self.afl_binary = afl_binary.result
+        self.dumped_memory = dumped_memory.result
+        self.inputs = inputs.result
+        self.outputs = outputs.result
+        self.harness_file = harness_file.result
 
+        try:
             global process
             process = subprocess.Popen([self.afl_binary, '-U' ,'-m' ,'none','-i', self.inputs, '-o', self.outputs, '--', 'python', self.harness_file, self.dumped_memory, '@@'],  preexec_fn=os.setsid)
             binja.log_info('Process {0} started'.format(os.getpgid(process.pid)))
-        except TypeError:
-            pass
+        except:
+            show_message_box("Afl-Unicorn", "Error please open git issue !",
+                             MessageBoxButtonSet.OKButtonSet, MessageBoxIcon.ErrorIcon)
         
     
     @classmethod
     def cancel_task(self, bv):
-        binja.log_info("Cancel process {0}".format(os.getpgid(process.pid)))
-        os.killpg(os.getpgid(process.pid), signal.SIGTERM) 
+        if(process):
+            binja.log_info("Cancel process {0}".format(os.getpgid(process.pid)))
+            os.killpg(os.getpgid(process.pid), signal.SIGTERM) 
 
 
     @classmethod
@@ -109,6 +115,9 @@ Attributes:
         self.end = 0
         self.avoid_addresses = []
         self.avoid = 0
+        self.dumped_memory = None
+        self.input_file = None
+        self.harness_file = None
 
     def set_start_address(self, bv, addr):
         if(addr in self.avoid_addresses):
@@ -273,25 +282,36 @@ Attributes:
         input_file.close()
     
     def test_harness(self, bv):
+        if(self.input_file != None):
+            print("TEST", self.input_file.result)
         separator = SeparatorField()
         dumped_memory = DirectoryNameField('Select folder with dumped memory')
         input_file = OpenFileNameField('Select input file')
         harness_file = OpenFileNameField('Select harness test file')
         get_form_input([separator, dumped_memory, input_file, harness_file], "Afl-unicorn Harness Test Menu")
+
+        if(dumped_memory.result == None or input_file.result == None or harness_file.result == None):
+            return
+
+        if(len(dumped_memory.result) <= 0 or len(input_file.result) <= 0 or len(harness_file.result) <=0):
+            show_message_box("Afl-Unicorn", "All fields are required !",
+                                            MessageBoxButtonSet.OKButtonSet, MessageBoxIcon.ErrorIcon)
+            return
         binja.log_info("Selected dumped memory folder: {0}".format(dumped_memory.result))
         binja.log_info("Selected input file: {0}".format(input_file.result))
         binja.log_info("Selected harness test file: {0}".format(harness_file.result))
+        
+        self.input_file = input_file
+        self.dumped_memory = dumped_memory
+        self.harness_file = harness_file
 
         try:
-            if(len(dumped_memory.result) <= 0 or len(input_file.result) <= 0 or len(harness_file.result) <=0):
-                show_message_box("Afl-Unicorn", "All fields are required !",
-                                                MessageBoxButtonSet.OKButtonSet, MessageBoxIcon.ErrorIcon)
-                return
-            output = subprocess.Popen(['python', harness_file.result, '-d', dumped_memory.result, input_file.result], stdout = subprocess.PIPE).communicate()[0]
+            output = subprocess.Popen(['python', self.harness_file.result, '-d', self.dumped_memory.result, self.input_file.result], stdout = subprocess.PIPE).communicate()[0]
             binja.log_info(output) 
-        except TypeError:
-            pass
 
+        except TypeError:
+              show_message_box("Afl-Unicorn", "Error please open git issue !",
+                             MessageBoxButtonSet.OKButtonSet, MessageBoxIcon.ErrorIcon)
 if __name__ == "__main__":
     pass
 else:
